+++ run-linux.sh
@@ -19,19 +19,21 @@ while [ "$process" = "0" ]
 Electric Book options
 ---------------------
 
-1. Create a print PDF
-2. Create a screen PDF
-3. Run as a website
-4. Create EPUB-ready files
-5. Export to Word
-6. Install or update dependencies
-7. Exit
+1  Create a print PDF
+2  Create a screen PDF
+3  Run as a website
+4  Create an epub
+5  Create an app
+6  Export to Word
+7  Convert source images to output formats
+8  Install or update dependencies
+x  Exit
 
 Enter a number and hit enter. "
 	read process
-	##################
-	# PRINT PDF      #
-	##################
+	#############
+	# PRINT PDF #
+	#############
 	if [ "$process" = "1" ]
 		then
 		# Remember the current folder
@@ -48,11 +50,10 @@ Enter a number and hit enter. "
 		echo "If you're outputting files in a subdirectory (e.g. a translation), type its name. Otherwise, hit enter. "
 		read printpdfsubdirectory
 		# Ask the user to add any extra Jekyll config files, e.g. _config.pdf-ebook.yml
-		echo -n "
-Any extra config files?
-Enter filenames (including any relative path), comma separated, no spaces. E.g.
-_configs/_config.myconfig.yml
-If not, just hit return."
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
 		read config
 		# Ask whether we're processing MathJax, to know whether to send the HTML via PhantomJS
 		echo "Does this book use MathJax? If no, hit enter. If yes, type any key then enter."
@@ -95,7 +96,7 @@ If not, just hit return."
 			fi
 			# Run prince, showing progress (-v), printing the docs in file-list
 			# and saving the resulting PDF to the _output folder
-			prince -v -l file-list -o ../../../_output/$printpdffilename.pdf
+			prince -v -l file-list -o ../../../_output/$printpdffilename.pdf --javascript
 			# Navigate back to where we began.
 			cd "$location"
 			# Tell the user we're done
@@ -103,7 +104,7 @@ If not, just hit return."
 			# Navigate to the _output folder...
 			cd _output
 			# and open the PDF we just created
-			# (for OSX, this will need to be open, not xdg-open)
+			# (for OSX, this is open, not xdg-open)
 			xdg-open $bookfolder.pdf
 			# Navigate back to where we began.
 			cd ../
@@ -114,9 +115,9 @@ If not, just hit return."
 		done
 		# Head back to the Electric Book options
 		process=0
-	##################
-	# SCREEN PDF     #
-	##################
+	##############
+	# SCREEN PDF #
+	##############
 	elif [ "$process" = 2 ]
 		then
 		# Remember the current folder
@@ -133,11 +134,10 @@ If not, just hit return."
 		echo "If you're outputting files in a subdirectory (e.g. a translation), type its name. Otherwise, hit enter. "
 		read screenpdfsubdirectory
 		# Ask the user to add any extra Jekyll config files, e.g. _config.pdf-ebook.yml
-		echo -n "
-Any extra config files?
-Enter filenames (including any relative path), comma separated, no spaces. E.g.
-_configs/_config.myconfig.yml
-If not, just hit return."
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
 		read config
 		# Ask whether we're processing MathJax, to know whether to send the HTML via PhantomJS
 		echo "Does this book use MathJax? If no, hit enter. If yes, type any key then enter."
@@ -180,7 +180,7 @@ If not, just hit return."
 			fi
 			# Run prince, showing progress (-v), printing the docs in file-list
 			# and saving the resulting PDF to the _output folder
-			prince -v -l file-list -o ../../../_output/$screenpdffilename.pdf
+			prince -v -l file-list -o ../../../_output/$screenpdffilename.pdf --javascript
 			# Navigate back to where we began.
 			cd "$location"
 			# Tell the user we're done
@@ -188,7 +188,7 @@ If not, just hit return."
 			# Navigate to the _output folder...
 			cd _output
 			# and open the PDF we just created
-			# (for OSX, this will need to be open, not xdg-open)
+			# (for OSX, this is open, not xdg-open)
 			xdg-open $bookfolder.pdf
 			# Navigate back to where we began.
 			cd ../
@@ -199,24 +199,22 @@ If not, just hit return."
 		done
 		# Head back to the Electric Book options
 		process=0
-	##################
-	# WEBSITE        #
-	##################
+	###########
+	# WEBSITE #
+	###########
 	elif [ "$process" = 3 ]
 		then
 		echo "Okay, let's make a website..."
 		# Ask the user to add any extra Jekyll config files, e.g. _config.pdf-ebook.yml
-		echo -n "
-Any extra config files?
-Enter filenames (including any relative path), comma separated, no spaces. E.g.
-_configs/_config.myconfig.yml
-If not, just hit return."
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
 		read config
 		# Ask the user to set a baseurl if needed
-		echo -n "Do you need a baseurl?
-If yes, enter it with no slashes at the start or end, e.g.
-my/base
-"
+		echo "Do you need a baseurl?"
+		echo "If yes, enter it with no slashes at the start or end, e.g."
+		echo "my/base"
 		read baseurl
 		# let the user know we're on it!
 		echo "Getting your site ready...
@@ -248,9 +246,9 @@ You may need to reload the web page once this server is running."
 		done
 		# Head back to the Electric Book options
 		process=0
-	##################
-	# EPUB           #
-	##################
+	########
+	# EPUB #
+	########
 	elif [ "$process" = 4 ]
 		then
 		# Remember the current folder
@@ -266,49 +264,146 @@ You may need to reload the web page once this server is running."
 		# Ask if we're outputting the files from a subdirectory (e.g. a translation)
 		echo "If you're outputting files in a subdirectory (e.g. a translation), type its name. Otherwise, hit enter. "
 		read epubsubdirectory
-		echo -n "What is the first file in your book? Usually the cover.
-(Don't include the file extension. Hit enter for the default '0-0-cover'.) "
-		read firstfile
-		if [ "$firstfile" = "" ]
-			then
-			firstfile="0-0-cover"
-		fi
-		# Ask the user to add any extra Jekyll config files, e.g. _config.myconfig.yml
-		echo -n "
-Any extra config files?
-Enter filenames (including any relative path), comma separated, no spaces. E.g.
-_configs/_config.myconfig.yml
-If not, just hit return."
-		read config
-		# Ask whether we're processing MathJax, to know whether to send the HTML via PhantomJS
-		echo "Does this book use MathJax? If no, hit enter. If yes, type any key then enter."
+		# Ask whether to keep the boilerplate epub mathjax directory
+		echo "Include mathjax? Enter y for yes (or enter for no)."
 		read epubmathjax
+		# Ask the user to add any extra Jekyll config files, e.g. _config.myconfig.yml
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
+		read config
 		# We're going to let users run this over and over by pressing enter
 		repeat=""
 		while [ "$repeat" = "" ]
 		do
 			# let the user know we're on it!
 			echo "Generating HTML..."
-			# ...and run Jekyll to build new HTML, enabling MathJax if necessary
-			if [ "$epubmathjax" = "" ]
-				then
-				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,$config"
-			else
+			# ...and run Jekyll to build new HTML
+			if [ $epubmathjax = "y" ]; then
 				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,_configs/_config.mathjax-enabled.yml,$config"
-			fi
-			# Navigate to the relevant text folder...
-			if [ "$epubsubdirectory" = "" ]
-				then
-				cd _site/$bookfolder/text
 			else
-				cd _site/$bookfolder/text/$epubsubdirectory
+				bundle exec jekyll build --config="_config.yml,_configs/_config.epub.yml,$config"
 			fi
-			# ...and open the first epub file there.
-			# Let the user know we're now going to open Sigil.
-			# NOTE: Sigil is not well supported on Linux, so many users won't have it.
-			echo "Opening Sigil..."
-			sigil "$firstfile.html"
-			# Open file browser to see epub-ready HTML files
+			# Now to assemble the epub
+			echo "Assembling epub..."
+			# Check if there are fonts to include
+			echo "Checking for fonts to include..."
+			epubfonts=""
+			countttf=`ls -1 fonts/*.ttf 2>/dev/null | wc -l`
+			if [ $countttf != 0 ]; then 
+				epubfonts="y"
+			fi
+			countotf=`ls -1 fonts/*.otf 2>/dev/null | wc -l`
+			if [ $countotf != 0 ]; then 
+				epubfonts="y"
+			fi
+			countwoff=`ls -1 fonts/*.woff 2>/dev/null | wc -l`
+			if [ $countwoff != 0 ]; then 
+				epubfonts="y"
+			fi
+			countwoff2=`ls -1 fonts/*.woff2 2>/dev/null | wc -l`
+			if [ $countwoff2 != 0 ]; then 
+				epubfonts="y"
+			fi
+			# Check if there are scripts to include
+			echo "Checking for scripts to include..."
+			epubscripts=""
+			countjs=`ls -1 js/*.js 2>/dev/null | wc -l`
+			if [ $countjs != 0 ]; then 
+				epubscripts="y"
+			fi
+			# Copy text (files in file-list only), images, fonts, styles and package.opf to epub
+			cd _site/"$bookfolder"
+			if [ "$epubsubdirectory" = "" ]; then
+				mkdir ../epub/text && cd text && cp `cat file-list` ../../epub/text/
+				cd ..
+			else
+				mkdir ../epub/text && cd $epubsubdirectory/text && cp `cat file-list` ../../../epub/text/
+				cd .. && cd ..
+			fi
+			if [ -d images ]; then
+				mkdir ../epub/images && cp -a images/. ../epub/images/
+			fi
+			if [ "$epubfonts" = "y" ]; then
+				mkdir ../epub/fonts && cp -a fonts/. ../epub/fonts/
+			fi
+			if [ -d styles ]; then
+				mkdir ../epub/styles && cp -a styles/. ../epub/styles/
+			fi
+			if [ "$epubmathjax" = "y" ]; then
+				mkdir ../epub/mathjax && cp -a ../assets/js/mathjax/. ../epub/mathjax/
+			fi
+			if [ "$epubscripts" = "y" ]; then
+				mkdir ../epub/js && cp -a js/. ../epub/js/
+			fi
+			if [ -e package.opf ]; then
+				cp package.opf ../epub/package.opf
+			fi
+			# First, though, if they exist, remove previous .zip and .epub files that we will replace.
+			echo "Removing previous zips or epubs..."
+			if [ -e "$location/_output/$bookfolder.zip" ]; then
+				rm "$location/_output/$bookfolder.zip"
+			fi
+			if [ -e "$location/_output/$bookfolder.epub" ]; then
+				rm "$location/_output/$bookfolder.epub"
+			fi
+			# Go into _site/epub to zip it to _output
+			cd ../epub
+			# First, though, remove the fonts folder if we dont' want it
+			if [ "$epubfonts" = "" ]; then
+				rm -r fonts
+			fi
+			# And remove the mathjax dir if we don't need it
+			if [ "$epubmathjax" = "" ]; then
+				rm -r mathjax
+			fi
+			# Now to compress the epub files
+			echo "Compressing epub..."
+			# Add the mimetype first, with no compression and no extra fields (-X)
+			zip --compression-method store -0 -X --quiet "$location/_output/$bookfolder.zip" mimetype
+			if [ -d images ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "images"
+			fi
+			if [ -d fonts ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "fonts"
+			fi
+			if [ -d styles ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "styles"
+			fi
+			if [ -d text ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "text"
+			fi
+			if [ -d mathjax ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "mathjax"
+			fi
+			if [ -d js ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" "js"
+			fi
+			if [ -d META-INF ]; then
+				zip --recurse-paths --quiet "$location/_output/$bookfolder.zip" META-INF
+			fi
+			if [ -e package.opf ]; then
+				zip --quiet "$location/_output/$bookfolder.zip" package.opf
+			fi
+			# Change file extension .zip to .epub
+			cd $location/_output
+			if [ -e "$bookfolder".zip ]; then
+				mv "$bookfolder".zip "$bookfolder".epub
+			fi
+			echo "Epub created!"
+			# Validation
+			echo "To run validation now, enter the path to the EpubCheck folder on your machine. E.g. /usr/bin/local/epubcheck-4.0.1"
+			echo "Or hit enter to skip EpubCheck validation."
+			echo "(You can get EpubCheck from https://github.com/IDPF/epubcheck/releases)"
+			read pathtoepubcheck
+			if [ "$pathtoepubcheck" = "" ]; then
+				echo "Okay, skipping EpubCheck. Try http://validator.idpf.org to validate separately."
+			else
+				java -jar "$pathtoepubcheck"/epubcheck.jar "$bookfolder".epub
+			fi
+			# Open file browser to see epub
+			# (for OSX, this is open, not xdg-open)
 			xdg-open .
 			# Navigate back to where we started
 			cd "$location"
@@ -319,10 +414,107 @@ If not, just hit return."
 		done
 		# Head back to the Electric Book options
 		process=0
+	###############
+	# CREATE APPS #
+	###############
+	elif [ "$process" = 5 ]
+		then
+		# Remember the current folder
+		location=$(pwd)
+		# Encouraging message
+		# (Building iOS not available on Linux, only newish Macs)
+		echo "Okay, let's make apps. First we'll generate the HTML, then we'll build Android and iOS apps."
+		echo "For the build, you need to have installed:"
+		echo "- Cordova"
+		echo "- Android Studio for the Android app"
+		echo "- XCode for the iOS app."
+		echo "Shall we build the apps, or just generate the HTML? Enter:"
+		echo "'a' to build only the Android app,"
+		echo "'i' to build only the iOs app,"
+		echo "'ai' to build both Android and iOS apps, or"
+		echo "'n' or just hit enter to skip building finished apps. "
+		appbuildgenerateapp=""
+		read appbuildgenerateapp
+		# Ask the user to add any extra Jekyll config files, e.g. _config.pdf-ebook.yml
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
+		read appconfig
+		# Ask whether we're enabling and including MathJax
+		echo "Do these books use MathJax? If yes, enter y. If no, hit enter."
+		read appmathjax
+		# We're going to let users run this over and over by pressing enter
+		repeat=""
+		while [ "$repeat" = "" ]
+		do
+			# let the user know we're on it!
+			echo "Generating HTML..."
+			# ...and run Jekyll to build new HTML
+			# with MathJax enabled if necessary
+			if [ "$appmathjax" = "" ]
+				then
+				bundle exec jekyll build --config="_config.yml,_configs/_config.app.yml,$config"
+			else
+				bundle exec jekyll build --config="_config.yml,_configs/_config.app.yml,_configs/_config.mathjax-enabled.yml,$config"
+			fi
+			# Put HTML into _site/app/www by moving everything in _site
+			# excluding the _site/app folder itself, into _site/app/www.
+			# (rsync lets us exclude, where cp does not)
+			# Suppress the console output.
+			echo "Copying files to app directory..."
+			mkdir _site/app/www
+			rsync -r _site/. _site/app/www --exclude="/app"
+			# Build the apps if required
+			if [ "$appbuildgenerateapp" = "a" ]
+				then
+				echo "Building Android app..."
+				cd _site/app
+				cordova build android
+				cd .. && cd ..
+				echo "Done. Opening folder containing Android app..."
+				# (On OSX, this will be open, not xdg-open.)
+				xdg-open _site/app/platforms/android/build/outputs/apk/
+			# Building iOS not available on Linux, only newish Macs
+			elif [ "$appbuildgenerateapp" = "i" ]
+				then
+				echo "Building iOS app..."
+				cd _site/app
+				cordova build ios
+				cd .. && cd ..
+				echo "Done. Opening folder containing iOS app..."
+				# (On OSX, this will be open, not xdg-open.)
+				xdg-open _site/app/platforms/ios/build/outputs/
+			elif [ "$appbuildgenerateapp" = "ai" ]
+				then
+				echo "Building Android app first, then iOS app."
+				echo "Building Android app..."
+				cordova build android
+				echo "Done. Opening folder containing Android app..."
+				# (On OSX, this will be open, not xdg-open.)
+				xdg-open _site/app/platforms/android/build/outputs/apk/
+				echo "Building iOS app..."
+				cordova build ios
+				echo "Done. Opening folder containing iOS app..."
+				# (On OSX, this will be open, not xdg-open.)
+				xdg-open _site/app/platforms/ios/build/outputs/
+			elif [ "$appbuildgenerateapp" = "" ] || [ "$appbuildgenerateapp" = "n" ]
+				then
+				echo "Skipping building apps. Opening app-ready files..."
+				# (On OSX, this will be open, not xdg-open.)
+				xdg-open _site
+			fi
+			# Ask the user if they want to rebuild
+			repeat=""
+			echo "Enter to run again, or any other key and enter to stop."
+			read repeat
+		done
+		# Head back to the Electric Book options
+		process=0
 	##################
 	# EXPORT TO WORD #
 	##################
-	elif [ "$process" = 5 ]
+	elif [ "$process" = 6 ]
 		then
 		echo "Okay, let's export to Word. You must have Pandoc installed for this to work."
 		# Remember current location
@@ -337,13 +529,11 @@ If not, just hit return."
 		echo "Okay, let's make Word files for $bookfolder..."
 		# Ask user which output format to work from
 		echo "Which format are we converting from? Enter a number or hit enter for the default 'print-pdf'. "
-		echo -n "
-1. Print PDF (default)
-2. Screen PDF
-3. Web
-4. Epub
-
-Enter a number and/or hit enter. "
+		echo "1. Print PDF (default)"
+		echo "2. Screen PDF"
+		echo "3. Web"
+		echo "4. Epub"
+		echo "Enter a number and/or hit enter. "
 		read fromformat
 		# Turn that choice into a variable named for the format
 		wordformatchoice=""
@@ -374,11 +564,10 @@ Enter a number and/or hit enter. "
 			fi
 		done
 		# Ask the user to add any extra Jekyll config files, e.g. _config.myconfig.yml
-		echo -n "
-Any extra config files?
-Enter filenames (including any relative path), comma separated, no spaces. E.g.
-_configs/_config.myconfig.yml
-If not, just hit return."
+		echo "Any extra config files?"
+		echo "Enter filenames (including any relative path), comma separated, no spaces. E.g."
+		echo "_configs/_config.myconfig.yml"
+		echo "If not, just hit return."
 		read config
 		# We're going to let users run this over and over by pressing enter
 		repeat=""
@@ -392,20 +581,18 @@ If not, just hit return."
 			cd _site/$bookfolder/text
 			# Update user
 			echo "Converting $bookfolder HTML to Word..."
+			# Before looping through file-list, remove blank lines.
+			# Why .bak? see https://stackoverflow.com/a/14570580/1781075
+			sed -i.bak '/^[[:space:]]*$/d' file-list
+			rm file-list.bak
 			# Loop through the list of files in file-list
 			# and convert them each from .html to .docx.
-			# We end up with the same filenames,
-			# with .docx extensions appended.
-# [Two loop methods to try here. This one:]
 			while read -r file
 			do
 				pandoc "$file" -f html -t docx -s -o $file.docx
 			done < file-list
-# [And this one:]
-#			for file in file-list
-#			do
-#				pandoc "$file" -f html -t docx -s -o $file.docx
-#			done
+			# We end up with the same filenames,
+			# with .docx extensions appended.
 			# Now we fix those file extensions
 			echo "Fixing file extensions..."
 			for file in *.html.docx
@@ -413,8 +600,9 @@ If not, just hit return."
 				mv "${file}" "${file/.html.docx/.docx}"
 			done
 			# Tell the user we're done
-			echo Done! Opening file explorer...
+			echo "Done! Opening file explorer..."
 			# Open file explorer to show the docx files
+			# (for OSX, this is open, not xdg-open)
 			xdg-open .
 			# Navigate back to where we began.
 			cd "$location"
@@ -426,22 +614,75 @@ If not, just hit return."
 		# Head back to the Electric Book options
 		process=0
 	##################
-	# INSTALL        #
+	# PROCESS IMAGES #
 	##################
-	elif [ "$process" = 6 ]
+	elif [ "$process" = 7 ]
 		then
-		echo "Running Bundler to update and install dependencies.
-If Bundler is not already installed, exit and run
-gem install bundler
-from the command line."
+	    echo "Let's convert your source images."
+	    echo "This process will optimise the images in a book's _source folder"
+	    echo "and copy them to the print-pdf, screen-pdf, web and epub image folders."
+	    echo "You need to have run 'Install or update dependencies' at least once,"
+	    echo "and have GraphicsMagick installed (http://www.graphicsmagick.org, or try"
+	    echo "brew install graphicsmagick"
+
+    	# Select which book to convert images for
+    	echo "Which book's images are you converting? Hit enter for the default 'book'."
+	    bookimagestoconvert=""
+	    read bookimagestoconvert
+	    if [ "$bookimagestoconvert" = "" ]
+	    	then
+	    	bookimagestoconvert="book"
+	    fi
+	    while [ ! -d $bookimagestoconvert ]
+	    do
+	    	echo "Sorry, there is no $bookimagestoconvert folder. Please try again."
+	    	read bookimagestoconvert
+	    done
+
+	    # Select whether we're converting images for a translation
+	    echo "Are we converting books in a translation? If not, hit enter."
+    	echo "Otherwise, enter the language code/translation directory name. "
+	    read convertimageslanguage
+
+	    # Only proceed if no language is set or the language folder exists
+	    while [[ ! -d "$bookimagestoconvert/$convertimageslanguage" && "$convertimageslanguage" != "" ]]
+	    do
+	    	echo "Sorry, there is no $bookimagestoconvert/$convertimageslanguage folder. Please try again."
+	    	read convertimageslanguage
+	    done
+
+		# We're going to let users run this over and over by pressing enter
+		repeat=""
+		while [ "$repeat" = "" ]
+		do
+		    # Run default gulp task
+		    gulp --book "$bookimagestoconvert" --language "$convertimageslanguage"
+
+			# Ask the user if they want to run that again
+			repeat=""
+			echo "Enter to run again, or any other key and enter to stop."
+			read repeat
+		done
+		# Head back to the Electric Book options
+		process=0
+
+	###########
+	# INSTALL #
+	###########
+	elif [ "$process" = 8 ]
+		then
+		echo "Running Bundler to update and install dependencies."
+		echo "If Bundler is not already installed, exit and run"
+		echo "gem install bundler"
+		echo "from the command line."
 		# Update gems
 		bundle update
 		# Install gems
 		bundler install
 		# Head back to the Electric Book options
 		process=0
-	##################
-	# EXIT           #
-	##################
+	########
+	# EXIT #
+	########
 	fi
 done